need to move the pre_option filter to the listener, so I changed the listener name to fit the use
