+++ includes/Services/InstaMigrateService.php
@@ -34,8 +34,6 @@ class InstaMigrateService {
 		$instawp_get_key_step = new GetInstaWpApiKey( $this->tracker );
 		$instawp_get_key_step->set_status( 'running' );
 		$this->insta_api_key = $instawp_get_key_step->get_api_key();
-
-		add_filter( 'pre_update_option_instawp_migration_details', array( $this, 'on_update_instawp_migration_details' ), 10, 2 );
 	}
 
 	/**
@@ -71,22 +69,5 @@ class InstaMigrateService {
 				array( 'status' => 400 )
 			);
 		}
-	}
-	/**
-	 * Listen instaWp option update to intercept the Push step and track it
-	 *
-	 * @param array $new_option status of migration
-	 * @param array $old_value previous status of migration
-	 * @return array
-	 */
-	public function on_update_instawp_migration_details( $new_option, $old_value ) {
-		if ( $old_value !== $new_option ) {
-			$mode   = isset( $new_option['mode'] ) ? $new_option['mode'] : '';
-			$status = isset( $new_option['status'] ) ? $new_option['status'] : '';
-			if ( 'push' === $mode && 'initiated' === $status ) {
-				$this->tracker->update_track( array( 'pushingStep' => array( 'status' => 'running' ) ) );
-			}
-		}
-		return $new_option;
 	}
 }
